requirements.txt
colorlog>=2.4.0
pbr>=0.10.0
six>=1.7.0
pyzmq>=14.3.1
msgpack-python>=0.4.2
